feat: add configurable eb-environment-names to generate-environment-info
Previously we were passing the prefix value in and configuring it from there. However, if we want to make changes to environments with rolling deploys, then we are going to need to use different environment names (e.g. replacing dev-api-docker with dev-api-docker-2). This problem then get's more complicated if we rely on convention when we only need to change one stage (e.g. only dev and qa but not production).

This change means we can simply say, here are the EB environment names for each stage and then we can access them switched based on the existing logic and keep our existing CD workflows intact.