feat: add production environment info back
With the refactor I had forgot to add the production semmantic regex check in. It was originally the output for all tags, but now it's explicitly the semmantic tags (checked via some regex). The code it's replacing was just a duplicated unrefactored line that was meant to be this change.