fix: remove test command to correctly check for re-tagging
I didn't understand correctly what [[ ... ]] meant. I now do. Basically it's the test command which evaluates a boolean expression and returns exit 0 if it's truthy and exit 1 if it's false. The if statement just evaluates something and continues if it was an exit 0, so that's how it ties together. The grep command isn't needing boolean evaluation, we are intending to rely on the grep -q exit code instead.