chore(check-oas-bundled): rename to check-oas-bundled from validate-oas
This was because I was going to do the CLI schema validation here. However that's proved difficult, instead I'm gonna run it into the swagger editor docker image as suggested (https://github.com/char0n/swagger-editor-validate) and that will validate it with output errors. The team were uncomfortable with using the HTTP approach.