feat(generate-environment-info): add ECS per-env inputs for deployments
This enables us to have specific values that we switch on for the ECS deployments. This is needed now our AWS environments are called `dev` and `qa` while our branching is `internal` and `staging` but also useful for when we need to make more changes to our ECS environments (e.g. changing task definition families etc).